requirements.txt
discord.py
gtts
PyNaCl
pymongo
python-dotenv
langdetect